Implemented a design pattern for the parser
Implemented strategy design pattern for the parser that works with multiple NOS
