+++ internal/parser/model_eos.go
@@ -1,0 +1,10 @@
+package parser
+
+import "fmt"
+
+type EOSParser struct {
+}
+
+func (p *EOSParser) PrintParserType() {
+   fmt.Println("Arista EOS")
+}
