+++ main.go
@@ -2,22 +2,29 @@ package main
 
 import (
 	"fmt"
+	"log"
 
 	"github.com/opennetworktools/NetConfigParser/internal/parser"
 	"github.com/opennetworktools/NetConfigParser/internal/utils"
 )
 
 func main() {
+	osType := "IOSXE"
 	path := "tests/configs/running-config-1.txt" // Path to config file
-	parser := parser.NewParser(path)
+	parser := parser.GetParser(osType, path)
+	if parser == nil {
+		log.Fatalf("Unsupported OS type: %s", osType)
+	}
 	parser.ParseConfig()
 
-	// preety-print
+	// pprint
 	// fmt.Printf("%+v\n", parser)
-
-	err := utils.WriteParserStructToJSON(parser, "out/parser.json")
+ 
+	err := utils.WriteParserStructToJSON(parser.GetConfigs(), "out/config.json")
 	if err != nil {
 		fmt.Println("Error creating JSON file:", err)
 	}
-	fmt.Println("Parser saved as JSON to parser.json")
-}
+	fmt.Println("Configs saved as JSON to configs.json")
+ }
+ 
+
